[misc] move pdf caching metrics into ContentCacheMetrics
